refactor(dftu): move DFT+U operator from module_operator_lcao to module_dftu
Move DFTU/OperatorDFTU<OperatorLCAO> operator implementation files
(dftu.hpp, dftu_lcao.h/cpp, dftu_fs.cpp, op_dftu_lcao.h/cpp) and the
related unit test (test_dftu.cpp -> dftu_lcao_test.cpp) from
module_operator_lcao to module_dftu. This is a step toward removing
module_operator_lcao and grouping DFT+U code by physical domain.

CMake changes:
- module_dftu CMakeLists adds the 3 cpp files guarded by ENABLE_LCAO,
  since module_dftu is unconditionally added by source_lcao while
  these files depend on operator_lcao.h
- module_operator_lcao CMakeLists drops the 3 cpp files
- source_lcao CMakeLists drops the 3 cpp entries from the hamilt_lcao
  OBJECT library (now provided by the dftu OBJECT library)
- New dftu_lcao_test target guarded by ENABLE_LCAO AND ENABLE_MPI
  matches the original MODULE_LCAO_operator_dftu_test build condition

Include path updates in 5 external call sites plus op_dftu_lcao.h
internal operator_lcao.h path (changed from same-directory relative
to full path since the file is no longer co-located).

Verified: cmake configure, make abacus_max_para + dftu_lcao_test,
and ctest dftu_pw_test/dftu_core_test/dftu_operator_test/dftu_lcao_test
all pass; agent_governance_check reports only include-path change
warnings with no errors.
