+++ source/source_lcao/module_dftu/test/CMakeLists.txt
@@ -17,3 +17,20 @@ AddTest(
   LIBS base device
   SOURCES dftu_operator_test.cpp
 )
+
+if(ENABLE_LCAO AND ENABLE_MPI)
+AddTest(
+  TARGET dftu_lcao_test
+  LIBS parameter psi base device container
+  SOURCES dftu_lcao_test.cpp ../dftu_lcao.cpp ../dftu_fs.cpp
+  ../../../source_pw/module_pwdft/dftu_base.cpp
+  ../../../source_hamilt/module_hcontainer/func_folding.cpp
+  ../../../source_hamilt/module_hcontainer/base_matrix.cpp
+  ../../../source_hamilt/module_hcontainer/hcontainer.cpp
+  ../../../source_hamilt/module_hcontainer/atom_pair.cpp
+  ../../../source_basis/module_ao/parallel_orbitals.cpp
+  ../../../source_basis/module_ao/orb_atomic_lm.cpp
+  ../../module_operator_lcao/test/tmp_mocks.cpp
+  ../../../source_hamilt/operator.cpp
+)
+endif()
