+++ source/source_lcao/module_operator_lcao/test/CMakeLists.txt
@@ -80,16 +80,6 @@ AddTest(
   tmp_mocks.cpp ../../../source_hamilt/operator.cpp
 )
 
-AddTest(
-  TARGET MODULE_LCAO_operator_dftu_test
-  LIBS parameter psi base device container
-  SOURCES test_dftu.cpp ../dftu_lcao.cpp ../dftu_fs.cpp ../../../source_pw/module_pwdft/dftu_base.cpp ../../../source_hamilt/module_hcontainer/func_folding.cpp
-  ../../../source_hamilt/module_hcontainer/base_matrix.cpp ../../../source_hamilt/module_hcontainer/hcontainer.cpp ../../../source_hamilt/module_hcontainer/atom_pair.cpp
-  ../../../source_basis/module_ao/parallel_orbitals.cpp
-  ../../../source_basis/module_ao/orb_atomic_lm.cpp
-  tmp_mocks.cpp ../../../source_hamilt/operator.cpp
-)
-
 install(FILES parallel_operator_tests.sh DESTINATION ${CMAKE_CURRENT_BINARY_DIR})
 find_program(BASH bash)
 add_test(NAME MODULE_LCAO_operators_para_test
